Move tests to api_view_test
