+++ courses/tests/api_view_test.py
@@ -1,10 +1,10 @@
 from rest_framework import status
 from rest_framework.test import APITestCase
-from rest_framework.test import APIRequestFactory
 from rest_framework.test import APIClient
 from rest_framework_simplejwt.tokens import RefreshToken
 
 from fixtures.factories.courses import LessonFactory
+from fixtures.factories.courses import SubmissionFactory
 from fixtures.factories.courses import TrackFactory
 from fixtures.factories.courses import UnitFactory
 from fixtures.factories.user import UserFactory
@@ -142,3 +142,29 @@ class AuthViewsTests(APITestCase):
         }
         response = self.admin_user_client.post(url, data, format='json')
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+
+class SumissionTests(APITestCase):
+
+    def setUp(self) -> None:
+        self.orig_submission = SubmissionFactory()
+        return super().setUp()
+    
+    def test_submission_permissions(self):
+        second_user = UserFactory()
+        refresh = RefreshToken.for_user(second_user)
+        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {refresh.access_token}')
+
+        url = f'/api/v1/submissions/{self.orig_submission.id}/'
+        data = {'submitted_code': 'print("Hi!")'}
+        response = self.client.patch(url, data)
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+        self.assertEqual(
+            response.json()['detail'],
+            'You do not have permission to perform this action.')
+        
+        orig_user = self.orig_submission.user
+        refresh = RefreshToken.for_user(orig_user)
+        self.client.credentials(HTTP_AUTHORIZATION=f'Bearer {refresh.access_token}')
+        response = self.client.patch(url, data)
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
